Split the local login and the main login into two seperate views.
